Fix/disk space error fix  (#1460)
* Update free-hosted-runner-disk-space action to version ac0009fd55b474dbf50864caa4f9da36a41e5ffe

* updating alfresco-build-tools to latest version


* Update free-hosted-runner-disk-space action to v18.23.0 and adjust swap size in scripts

* Add permissions for contents and pull-requests in CI configuration

* Add permissions for contents and pull-requests in CI configuration

* removed pmd scan permission block

---------

+++ scripts/ci/create_swap.sh
@@ -5,8 +5,12 @@ SIZE_IN_GB="${1}"
 
 sudo free -m -t
 sudo swapoff -a
-sudo fallocate -l ${SIZE_IN_GB}G /swapfile
-sudo dd if=/dev/zero of=/swapfile bs=1024 count=$((${SIZE_IN_GB} * 1024 * 1024))
+sudo rm -f /swapfile
+if command -v fallocate >/dev/null 2>&1; then
+  sudo fallocate -l ${SIZE_IN_GB}G /swapfile
+else
+  sudo dd if=/dev/zero of=/swapfile bs=1024 count=$((${SIZE_IN_GB} * 1024 * 1024))
+fi
 sudo chmod 600 /swapfile
 sudo mkswap /swapfile
 sudo swapon /swapfile
